job-agent-openai/requirements.txt
# Streamlit UI
streamlit==1.28.0

# OpenAI API
openai==1.27.0

# Async HTTP client
aiohttp==3.8.5

# Data handling & computation
numpy==1.26.0
pandas==2.1.0

# Load environment variables (API keys etc.)
python-dotenv==1.0.0

# Optional: for more advanced embeddings / ML
scikit-learn==1.3.2

# Optional: for pretty printing / structured logs
rich==13.6.0